+++ smtk/attribute/ValueItemDefinition.cxx
@@ -131,7 +131,7 @@ ValueItemDefinition::buildExpressionItem(ValueItem *vitem, int position) const
     smtk::dynamic_pointer_cast<smtk::attribute::RefItem>
     (this->m_expressionDefinition->buildItem(vitem, position, -1));
   aref->setDefinition(this->m_expressionDefinition);
-  vitem->m_expressions[position] = aref;
+  vitem->m_expressions[static_cast<size_t>(position)] = aref;
 }
 //----------------------------------------------------------------------------
 void
